Fix Eigen3 Dependencies in CMakeLists file
Add Eigen3 as target dependency for TeleporIngestorCommon and move
template functions from utils to TeleportIngestorCommon.
