+++ rmf_demo_plugins/rmf_plugins_common/include/rmf_plugins_common/ingestor_common.hpp
@@ -21,6 +21,7 @@
 #include <string>
 #include <unordered_map>
 
+#include <Eigen/Geometry>
 #include <rclcpp/rclcpp.hpp>
 
 #include <rmf_fleet_msgs/msg/fleet_state.hpp>
@@ -95,6 +96,62 @@ private:
     std::function<void()> transport_model_cb,
     const std::string& fleet_name);
 };
+
+// Ignition math version agnostic conversion functions
+template<typename IgnQuatT>
+inline void convert(const Eigen::Quaterniond& _q, IgnQuatT& quat)
+{
+  quat.W() = _q.w();
+  quat.X() = _q.x();
+  quat.Y() = _q.y();
+  quat.Z() = _q.z();
+}
+
+template<typename IgnVec3T>
+inline void convert(const Eigen::Vector3d& _v, IgnVec3T& vec)
+{
+  vec.X() = _v[0];
+  vec.Y() = _v[1];
+  vec.Z() = _v[2];
+}
+
+template<typename IgnVec3T>
+inline Eigen::Vector3d convert_vec(const IgnVec3T& _v)
+{
+  return Eigen::Vector3d(_v[0], _v[1], _v[2]);
+}
+
+template<typename IgnQuatT>
+inline Eigen::Quaterniond convert_quat(const IgnQuatT& _q)
+{
+  Eigen::Quaterniond quat;
+  quat.w() = _q.W();
+  quat.x() = _q.X();
+  quat.y() = _q.Y();
+  quat.z() = _q.Z();
+
+  return quat;
+}
+
+template<typename IgnPoseT>
+inline auto convert_to_pose(const Eigen::Isometry3d& _tf)
+{
+  IgnPoseT pose;
+  convert(Eigen::Vector3d(_tf.translation()), pose.Pos());
+  convert(Eigen::Quaterniond(_tf.linear()), pose.Rot());
+
+  return pose;
+}
+
+template<typename IgnPoseT>
+inline Eigen::Isometry3d convert_pose(const IgnPoseT& _pose)
+{
+  Eigen::Isometry3d tf = Eigen::Isometry3d::Identity();
+  tf.translation() = convert_vec(_pose.Pos());
+  tf.linear() = Eigen::Matrix3d(convert_quat(_pose.Rot()));
+
+  return tf;
+}
 
 } // namespace rmf_ingestor_common
